get success jwt token

+++ src/main/java/com/green/flighter/repository/TicketRepository.java
@@ -3,5 +3,5 @@ package com.green.flighter.repository;
 import com.green.flighter.model.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 }
